+++ src/main/java/com/example/modeladov1/model/Tienda.java
@@ -28,7 +28,7 @@ public class Tienda {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_tienda")
-    private Long id_tienda;
+    private int id_tienda;
 
     @Column(name="descripcion")
     private String descripcion;
@@ -37,11 +37,11 @@ public class Tienda {
 
     // Getters y setters
 
-    public Long getId_tienda() {
+    public int getId_tienda() {
         return id_tienda;
     }
 
-    public void setId_tienda(Long id_tienda) {
+    public void setId_tienda(int id_tienda) {
         this.id_tienda = id_tienda;
     }
